+++ src/main/java/tfcflorae/world/surface/builder/MangroveSurfaceBuilder.java
@@ -34,7 +34,7 @@ public class MangroveSurfaceBuilder implements SurfaceBuilder
 
     private void buildMudSurface(SurfaceBuilderContext context, int startY, int endY)
     {
-        final float noise = surfaceMaterialNoise.noise(context.pos().getX(), context.pos().getZ()) * 0.9f + context.random().nextFloat() * 0.1f;
+        final double noise = surfaceMaterialNoise.noise(context.pos().getX(), context.pos().getZ()) * 0.9f + context.random().nextFloat() * 0.1f;
         NormalSurfaceBuilder.INSTANCE.buildSurface(context, startY, endY, noise < 0f ? SurfaceStates.GRASS : SurfaceStates.MUD, SurfaceStates.GRAVEL, SurfaceStates.SANDSTONE_OR_GRAVEL);
     }
 }
