+++ src/main/java/tfcflorae/world/surface/builder/RainbowMountainsSurfaceBuilder.java
@@ -53,7 +53,7 @@ public class RainbowMountainsSurfaceBuilder implements SurfaceBuilder
 
         if (heightNoise > 118 + gaussSpread)
         {
-            float surfaceMaterialValue = surfaceMaterialNoise.noise(context.pos().getX(), context.pos().getZ()) + 0.1f * context.random().nextFloat() - 0.05f;
+            double surfaceMaterialValue = surfaceMaterialNoise.noise(context.pos().getX(), context.pos().getZ()) + 0.1f * context.random().nextFloat() - 0.05f;
             if (surfaceMaterialValue > 0.3f)
             {
                 surfaceBuilder.buildSurface(context, startY, endY, SurfaceStates.COBBLE, SurfaceStates.COBBLE, SurfaceStates.RAW);
